+++ src/full/Agda/TypeChecking/Unquote.hs
@@ -29,11 +29,16 @@ import qualified Agda.Syntax.Reflected as R
 import qualified Agda.Syntax.Abstract as A
 import Agda.Syntax.Abstract.Views
 import Agda.Syntax.Translation.InternalToAbstract
+import Agda.Syntax.Translation.ConcreteToAbstract
 import Agda.Syntax.Literal
+import qualified Agda.Syntax.Concrete as C
+import Agda.Syntax.Concrete.Name (simpleName)
 import Agda.Syntax.Position
 import Agda.Syntax.Info as Info
 import Agda.Syntax.Translation.ReflectedToAbstract
-import Agda.Syntax.Scope.Base (KindOfName(ConName, DataName))
+import Agda.Syntax.Scope.Base (KindOfName(ConName, DataName)
+                              , scopeLocals, LocalVar(LocalVar), BindingSource(MacroBound) )
+import Agda.Syntax.Parser
 
 import Agda.Interaction.Library ( ExeName )
 import Agda.Interaction.Options ( optTrustedExecutables, optAllowExec )
@@ -53,6 +58,7 @@ import Agda.TypeChecking.Primitive
 import Agda.TypeChecking.ReconstructParameters
 import Agda.TypeChecking.CheckInternal
 import Agda.TypeChecking.InstanceArguments
+import Agda.TypeChecking.Warnings
 
 import {-# SOURCE #-} Agda.TypeChecking.Rules.Term
 import {-# SOURCE #-} Agda.TypeChecking.Rules.Def
@@ -615,6 +621,7 @@ evalTCM v = Bench.billTo [Bench.Typing, Bench.Reflection] do
              , (f `isDef` getBuiltin' builtinAgdaTCMDefineFun,  uqFun2 tcDefineFun  u v)
              , (f `isDef` getBuiltin' builtinAgdaTCMQuoteOmegaTerm, tcQuoteTerm (sort $ Inf UType 0) (unElim v))
              , (f `isDef` getBuiltin' builtinAgdaTCMPragmaForeign, tcFun2 tcPragmaForeign u v)
+             , (f `isDef` getBuiltin' builtinAgdaTCMCheckFromString, tcFun2 tcCheckFromString u v)
              ]
              failEval
     I.Def f [l, a, u] ->
@@ -786,6 +793,23 @@ evalTCM v = Bench.billTo [Bench.Typing, Bench.Reflection] do
       else
         quoteTerm =<< process v
 
+
+    tcCheckFromString :: Text -> R.Type -> TCM Term
+    tcCheckFromString str a = do
+      (C.ExprWhere c wh , _) <- runPM $ parsePosString exprWhereParser (startPos Nothing) (T.unpack str)
+      r <- isReconstructed
+      e <- concreteToAbstract_ c
+      a <- workOnTypes $ locallyReduceAllDefs $ isType_ =<< toAbstract_ a
+
+      v <- checkExpr e a
+      if r then do
+        v <- process v
+        v <- locallyReduceAllDefs $ reconstructParameters a v
+        locallyReconstructed (quoteTerm v)
+      else
+        quoteTerm =<< process v
+
+
     tcQuoteTerm :: Type -> Term -> UnquoteM Term
     tcQuoteTerm a v = liftTCM $ do
       r <- isReconstructed
@@ -851,9 +875,11 @@ evalTCM v = Bench.billTo [Bench.Typing, Bench.Reflection] do
         quoteDomWithName (x, t) = toTerm <*> pure (T.pack x, t)
 
     extendCxt :: Text -> Arg R.Type -> UnquoteM a -> UnquoteM a
-    extendCxt s a m = do
+    extendCxt s' a m = withFreshName noRange (T.unpack s') $ \s -> do
       a <- workOnTypes $ locallyReduceAllDefs $ liftTCM $ traverse (isType_ <=< toAbstract_) a
-      liftU1 (addContext (s, domFromArg a :: Dom Type)) m
+
+      locallyScope scopeLocals ((simpleName (T.unpack s') , LocalVar s MacroBound []) :)
+          $ liftU1 (addContext (s, domFromArg a :: Dom Type)) m
 
     tcExtendContext :: Term -> Term -> Term -> UnquoteM Term
     tcExtendContext s a m = do